Move shaders to `shaders` module
