+++ src/lib.rs
@@ -20,3 +20,4 @@ pub use glyph::GlyphCache;
 
 mod back_end;
 mod glyph;
+mod shaders;
